requirements.txt
Flask==2.3.3
Pillow==11.1.0
numpy==1.26.4
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.0